+++ src/main/java/controller/EnterPlayerPhase.java
@@ -1,0 +1,24 @@
+package controller;
+
+import java.util.List;
+import model.Player;
+import view.InputView;
+
+public class EnterPlayerPhase implements GamePhase {
+
+    public EnterPlayerPhase() {
+    }
+
+    @Override
+    public void execute(GameContext gameContext) {
+        List<String> inputNames = InputView.readPlayerNames();
+        List<Player> players = createPlayers(inputNames);
+        gameContext.setPlayers(players);
+    }
+
+    private List<Player> createPlayers(List<String> names) {
+        return names.stream()
+                .map(Player::new)
+                .toList();
+    }
+}
