+++ src/main/java/controller/GamePhase.java
@@ -1,0 +1,6 @@
+package controller;
+
+public interface GamePhase {
+
+    void execute(GameContext gameContext);
+}
